Add country support for NTFEmail templates
